+++ src/sync/value_initializer.rs
@@ -1,24 +1,26 @@
 use parking_lot::RwLock;
 use std::{
+    any::{Any, TypeId},
     error::Error,
     hash::{BuildHasher, Hash},
     sync::Arc,
 };
 
-type Waiter<V> = Arc<RwLock<Option<Result<V, Arc<Box<dyn Error + Send + Sync + 'static>>>>>>;
+type ErrorObject = Arc<dyn Any + Send + Sync + 'static>;
+type Waiter<V> = Arc<RwLock<Option<Result<V, ErrorObject>>>>;
 
-#[allow(clippy::redundant_allocation)]
-// https://rust-lang.github.io/rust-clippy/master/index.html#redundant_allocation
-pub(crate) enum InitResult<V> {
+pub(crate) enum InitResult<V, E> {
     Initialized(V),
     ReadExisting(V),
-    // This `Arc<Box<dyn ..>>` creates an extra heap allocation. This will be
-    // addressed by Moka v0.6.0.
-    InitErr(Arc<Box<dyn Error + Send + Sync + 'static>>),
+    InitErr(Arc<E>),
 }
 
 pub(crate) struct ValueInitializer<K, V, S> {
-    waiters: moka_cht::SegmentedHashMap<Arc<K>, Waiter<V>, S>,
+    // TypeId is the type ID of the concrete error type of generic type E in
+    // try_init_or_read(). We use the type ID as a part of the key to ensure that
+    // we can always downcast the trait object ErrorObject (in Waiter<V>) into
+    // its concrete type.
+    waiters: moka_cht::SegmentedHashMap<(Arc<K>, TypeId), Waiter<V>, S>,
 }
 
 impl<K, V, S> ValueInitializer<K, V, S>
@@ -33,13 +35,13 @@ where
         }
     }
 
-    pub(crate) fn init_or_read(&self, key: Arc<K>, init: impl FnOnce() -> V) -> InitResult<V> {
+    pub(crate) fn init_or_read(&self, key: Arc<K>, init: impl FnOnce() -> V) -> InitResult<V, ()> {
         use InitResult::*;
 
         let waiter = Arc::new(RwLock::new(None));
         let mut lock = waiter.write();
 
-        match self.try_insert_waiter(&key, &waiter) {
+        match self.try_insert_waiter(&key, TypeId::of::<()>(), &waiter) {
             None => {
                 // Inserted. Evaluate the init closure.
                 let value = init();
@@ -58,16 +60,18 @@ where
         }
     }
 
-    pub(crate) fn try_init_or_read<F>(&self, key: Arc<K>, init: F) -> InitResult<V>
+    pub(crate) fn try_init_or_read<F, E>(&self, key: Arc<K>, init: F) -> InitResult<V, E>
     where
-        F: FnOnce() -> Result<V, Box<dyn Error + Send + Sync + 'static>>,
+        F: FnOnce() -> Result<V, E>,
+        E: Error + Send + Sync + 'static,
     {
         use InitResult::*;
 
+        let type_id = TypeId::of::<E>();
         let waiter = Arc::new(RwLock::new(None));
         let mut lock = waiter.write();
 
-        match self.try_insert_waiter(&key, &waiter) {
+        match self.try_insert_waiter(&key, type_id, &waiter) {
             None => {
                 // Inserted. Evaluate the init closure.
                 match init() {
@@ -76,10 +80,10 @@ where
                         Initialized(value)
                     }
                     Err(e) => {
-                        let err = Arc::new(e);
+                        let err: ErrorObject = Arc::new(e);
                         *lock = Some(Err(Arc::clone(&err)));
-                        self.remove_waiter(&key);
-                        InitErr(err)
+                        self.remove_waiter(&key, type_id);
+                        InitErr(err.downcast().unwrap())
                     }
                 }
             }
@@ -89,7 +93,7 @@ where
                 std::mem::drop(lock);
                 match &*res.read() {
                     Some(Ok(value)) => ReadExisting(value.clone()),
-                    Some(Err(e)) => InitErr(Arc::clone(e)),
+                    Some(Err(e)) => InitErr(Arc::clone(e).downcast().unwrap()),
                     None => unreachable!(),
                 }
             }
@@ -97,15 +101,21 @@ where
     }
 
     #[inline]
-    pub(crate) fn remove_waiter(&self, key: &Arc<K>) {
-        self.waiters.remove(key);
+    pub(crate) fn remove_waiter(&self, key: &Arc<K>, type_id: TypeId) {
+        let key = Arc::clone(key);
+        self.waiters.remove(&(key, type_id));
     }
 
-    fn try_insert_waiter(&self, key: &Arc<K>, waiter: &Waiter<V>) -> Option<Waiter<V>> {
+    fn try_insert_waiter(
+        &self,
+        key: &Arc<K>,
+        type_id: TypeId,
+        waiter: &Waiter<V>,
+    ) -> Option<Waiter<V>> {
         let key = Arc::clone(key);
         let waiter = Arc::clone(waiter);
 
         self.waiters
-            .insert_with_or_modify(key, || waiter, |_, w| Arc::clone(w))
+            .insert_with_or_modify((key, type_id), || waiter, |_, w| Arc::clone(w))
     }
 }